Switch export to float16 compute and reduce context to 5k
- _rmsnorm/_rmsnorm_noscale: cast output to float16 (internals stay
  float32 for numerical stability). Previously returning float32
  propagated to ALL downstream matmuls — CoreML decompressed int4
  weights into float32 buffers (~18 GB) instead of float16 (~9 GB).
- Softmax: cast output to float16 after float32 computation. Attention
  probabilities are in [0,1], perfectly representable in float16.
  This makes ALL matmuls float16 except the final logits dot product.
- compute_precision: change from FLOAT32 to FLOAT16 to match the
  now-float16 graph. add_fp16_cast pass is still removed to protect
  the intentional float32 softmax and RMSNorm internals.
- MAX_SEQ_LEN: 10000 → 5000 for iPhone-appropriate context length.
  Halves global KV cache memory (3 caches × head_dim=512).

Requires re-export: uv run gemma-export

+++ gemma_chat/decode_coreml.py
@@ -77,13 +77,13 @@ def empty_pos_ring(cfg: Gemma4Config = E2B_CONFIG) -> jnp.ndarray:
 def _rmsnorm(x, scale):
     x32 = x.astype(jnp.float32)
     var = jnp.mean(jnp.square(x32), axis=-1, keepdims=True)
-    return x32 * jax.lax.rsqrt(var + 1e-6) * scale.astype(jnp.float32)
+    return (x32 * jax.lax.rsqrt(var + 1e-6) * scale.astype(jnp.float32)).astype(jnp.float16)
 
 
 def _rmsnorm_noscale(x):
     x32 = x.astype(jnp.float32)
     var = jnp.mean(jnp.square(x32), axis=-1, keepdims=True)
-    return x32 * jax.lax.rsqrt(var + 1e-6)
+    return (x32 * jax.lax.rsqrt(var + 1e-6)).astype(jnp.float16)
 
 
 def _ple_for_tokens(params, token_ids, cfg: Gemma4Config):
@@ -211,7 +211,7 @@ def _attn_decode(lp, x, position, cfg: Gemma4Config, attn_type: str,
 
     w = jnp.matmul(qt, jnp.swapaxes(kt, -2, -1))           # (1, H, 1, max_len)
     w = jnp.where(valid[jnp.newaxis, jnp.newaxis, jnp.newaxis], w, -10000.0)
-    w = jax.nn.softmax(w.astype(jnp.float32), axis=-1)
+    w = jax.nn.softmax(w.astype(jnp.float32), axis=-1).astype(jnp.float16)
 
     out = jnp.matmul(w, vt)                                 # (1, H, 1, hd)
     out = jnp.transpose(out, (0, 2, 1, 3)).reshape(1, 1, num_heads * hd)
@@ -414,7 +414,7 @@ def _attn_chunk(lp, x, positions, start_pos, cfg: Gemma4Config, attn_type: str,
         mask = pos_k[jnp.newaxis, :] <= pos_q[:, jnp.newaxis]  # (C, max_len)
 
     w = jnp.where(mask[jnp.newaxis, jnp.newaxis], w, -10000.0)
-    w = jax.nn.softmax(w.astype(jnp.float32), axis=-1)
+    w = jax.nn.softmax(w.astype(jnp.float32), axis=-1).astype(jnp.float16)
 
     out = jnp.matmul(w, vt)                                    # (1, H, C, hd)
     out = jnp.transpose(out, (0, 2, 1, 3)).reshape(1, C, num_heads * hd)
@@ -465,7 +465,7 @@ def _attn_chunk_shared(lp, x, positions, start_pos, cfg: Gemma4Config, attn_type
         mask = pos_k[jnp.newaxis, :] <= pos_q[:, jnp.newaxis]
 
     w = jnp.where(mask[jnp.newaxis, jnp.newaxis], w, -10000.0)
-    w = jax.nn.softmax(w.astype(jnp.float32), axis=-1)
+    w = jax.nn.softmax(w.astype(jnp.float32), axis=-1).astype(jnp.float16)
 
     out = jnp.transpose(jnp.matmul(w, vt), (0, 2, 1, 3)).reshape(1, C, num_heads * hd)
     return jnp.dot(out, sa['o_proj']['kernel'])
